Subindo sprint 3 para produção

+++ src/controllers/produtos.js
@@ -1,21 +1,28 @@
-const { verificaCategoria } = require("../services/categorias");
-const { insertProduto,
+const path = require("path");
+const { randomUUID } = require('crypto');
+const { verificarCategoriaService } = require("../services/categorias");
+const { cadastrarProdutoService,
     obterListaProdutos,
     atualizarProdutoService,
-    obterProdutoPorId, excluirProdutoService
+    obterProdutoPorId,
+    excluirProdutoService,
+    enviarImagem,
+    deletarImagem,
+    atualizarImagemService,
+    verificarSeExistePedidoParaProduto
 } = require("../services/produtos");
 
-const cadastrarProduto = async (req, res) => {
+const cadastrarProdutos = async (req, res) => {
     const { descricao, quantidade_estoque, valor, categoria_id } = req.body;
 
     try {
-        const categoriaEncontrada = await verificaCategoria(categoria_id);
+        const categoriaEncontrada = await verificarCategoriaService(categoria_id);
 
         if (!categoriaEncontrada) {
             return res.status(400).json({ mensagem: "Categoria informada não encontrada." });
         }
 
-        const produtoCadastrado = await insertProduto(descricao, quantidade_estoque, valor, categoria_id);
+        const produtoCadastrado = await cadastrarProdutoService(descricao, quantidade_estoque, valor, categoria_id);
 
         return res.status(201).json(produtoCadastrado);
 
@@ -28,7 +35,7 @@ const listarProdutos = async (req, res) => {
     const categoria_id = req.query.categoria_id;
 
     try {
-        if (categoria_id && !(await verificaCategoria(categoria_id))) {
+        if (categoria_id && !(await verificarCategoriaService(categoria_id))) {
             return res.status(400).json({ mensagem: "A categoria informada não foi encontrada" });
         }
 
@@ -69,7 +76,7 @@ const atualizarProduto = async (req, res) => {
             return res.status(404).json({ mensagem: "Produto não encontrado." });
         }
 
-        const categoriaEncontrada = await verificaCategoria(categoria_id);
+        const categoriaEncontrada = await verificarCategoriaService(categoria_id);
 
         if (!categoriaEncontrada) {
             return res.status(400).json({ mensagem: "Categoria informada não encontrada." });
@@ -94,7 +101,16 @@ const excluirProduto = async (req, res) => {
             return res.status(404).json({ mensagem: "Produto não encontrado." });
         }
 
+        const existeProdutoPedido = await verificarSeExistePedidoParaProduto(id);
+
+        if (existeProdutoPedido) {
+            return res.status(404).json({ mensagem: "Não é possível excluir o produto, pois ele está vinculado a um ou mais pedidos." });
+        }
+
         await excluirProdutoService(id);
+
+        const pathImagemProdutoDeletado = produtoExistente.imagem_url.slice(70);
+        await deletarImagem(pathImagemProdutoDeletado);
 
         return res.status(200).json({ mensagem: "Produto excluído com sucesso." });
 
@@ -103,10 +119,61 @@ const excluirProduto = async (req, res) => {
     }
 }
 
+const adicionarImagemProduto = async (req, res) => {
+    const { id } = req.params;
+
+    try {
+        const produtoExistente = await obterProdutoPorId(id);
+
+        if (!produtoExistente) {
+            return res.status(404).json({ mensagem: "Produto não encontrado." });
+        }
+
+        if (!req.file) {
+            if (!produtoExistente.imagem_url) {
+                return res.status(401).json({ mensagem: "Produto não contém imagem." });
+            }
+
+            const pathImagemProdutoDeletada = produtoExistente.imagem_url.slice(70);
+            await deletarImagem(pathImagemProdutoDeletada);
+
+            await atualizarImagemService(id, null);
+
+            return res.status(200).json({ mensagem: "Imagem excluida com sucesso" });
+        }
+
+        const { originalname, buffer, mimetype } = req.file
+
+        if (produtoExistente.imagem_url !== "" && produtoExistente.imagem_url !== null) {
+            const pathImagemProdutoDeletada = produtoExistente.imagem_url.slice(70);
+            await deletarImagem(pathImagemProdutoDeletada);
+        }
+
+        const nomeArquivo = `${randomUUID()}${path.extname(originalname)}`
+        const imagem = await enviarImagem(
+            `imagens/${nomeArquivo}`,
+            buffer,
+            mimetype
+        )
+        const url = `${process.env.STORAGE_BASEURL}/${process.env.STORAGE_BUCKET}/imagens/${nomeArquivo}`;
+
+        await atualizarImagemService(id, url);
+
+        return res.status(201).json({
+            path: `imagens/${nomeArquivo}`,
+            url
+        })
+
+    } catch (error) {
+        return res.status(500).json({ mensagem: "Erro interno do servidor." });
+    }
+}
+
 module.exports = {
-    cadastrarProduto,
+    cadastrarProdutos,
     listarProdutos,
     detalharProduto,
     atualizarProduto,
-    excluirProduto
+    excluirProduto,
+    adicionarImagemProduto
 }
